try to tidy up auth.Server.generateCerts a bit

+++ lib/auth/join.go
@@ -143,12 +143,20 @@ func (a *Server) generateCerts(ctx context.Context, provisionToken types.Provisi
 
 		joinMethod := provisionToken.GetJoinMethod()
 
-		// certs for IAM method should not be renewable
+		// Static tokens should produce renewable certificates, as the bot has
+		// no other way to prove its identity after initial joining. Dynamic
+		// token types (IAM, OIDC-GCP) should not be renewable, we will expect
+		// the bot to continually "join" using fresh proof from the third party
+		// authority.
 		var renewable bool
+		// Mark ephemeral token types for deletion after use.
+		var shouldDeleteToken bool
 		switch joinMethod {
 		case types.JoinMethodToken:
 			renewable = true
-		case types.JoinMethodIAM, types.JoinMethodOIDCGCP:
+			shouldDeleteToken = true
+		case types.JoinMethodIAM,
+			types.JoinMethodOIDCGCP:
 			renewable = false
 		default:
 			return nil, trace.BadParameter("unsupported join method %q for bot", joinMethod)
@@ -160,17 +168,11 @@ func (a *Server) generateCerts(ctx context.Context, provisionToken types.Provisi
 			return nil, trace.Wrap(err)
 		}
 
-		switch joinMethod {
-		case types.JoinMethodToken:
-			// delete ephemeral bot join tokens so they can't be re-used
+		if shouldDeleteToken {
 			if err := a.DeleteToken(ctx, provisionToken.GetName()); err != nil {
 				log.WithError(err).Warnf("Could not delete bot provision token %q after generating certs",
 					string(backend.MaskKeyName(provisionToken.GetName())))
 			}
-		case types.JoinMethodIAM, types.JoinMethodOIDCGCP:
-			// don't delete long-lived IAM/OIDC join tokens
-		default:
-			return nil, trace.BadParameter("unsupported join method %q for bot", joinMethod)
 		}
 
 		log.Infof("Bot %q has joined the cluster.", botName)
